refactor: 💡 add useSignUp which is hooks to create account
